+++ src/api/groups.rs
@@ -16,6 +16,7 @@ mod groups;
 pub mod members;
 pub mod milestones;
 pub mod projects;
+pub mod subgroups;
 
 pub use create::BranchProtection;
 pub use create::CreateGroup;